Restylize blockquotes.
In the original version of this template, blockquotes were obviously thought of
as something composed of only a few words, perhaps a sentence.  My quotes tend
to be more... blocky, however, and I don't particularly want a bunch of 48px
text filling up the page.

Since I made the text smaller, I decided it needed something else to mark it
off.  So, we've got raquo and some light blue background color, themed off the
link color.